updated pr comments
