+++ src/scripts/run-production-migration.js
@@ -8,145 +8,6 @@
 require('dotenv').config({ path: '../.env' })
 const MentoringDataMigrator = require('./mentoring-data-migration')
 const readline = require('readline')
-const { Sequelize } = require('sequelize')
-
-/**
- * Validate that all organization_ids in database tables exist in the CSV file
- * Fails the migration if any database organization_ids are missing from CSV
- */
-async function validateDatabaseOrgsCoveredByCSV(orgLookupCache, sequelize) {
-	console.log('\n🔍 Validating database organization_ids coverage in CSV...')
-
-	const missingOrgs = new Set()
-	const csvOrgIds = new Set(orgLookupCache.keys())
-
-	// Get all organization_ids from tables that have organization_id column
-	const tablesWithOrgId = [
-		'availabilities',
-		'default_rules',
-		'entity_types',
-		'file_uploads',
-		'forms',
-		'notification_templates',
-		'organization_extension',
-		'report_queries',
-		'report_role_mapping',
-		'report_types',
-		'reports',
-		'resources',
-		'role_extensions',
-		'sessions',
-		'user_extensions',
-	]
-
-	for (const tableName of tablesWithOrgId) {
-		try {
-			// Check if table exists first
-			const tableExists = await sequelize.query(
-				`SELECT EXISTS (SELECT FROM information_schema.tables WHERE table_name = '${tableName}')`,
-				{ type: Sequelize.QueryTypes.SELECT }
-			)
-
-			if (!tableExists[0].exists) {
-				console.log(`⚠️  Table ${tableName} does not exist, skipping`)
-				continue
-			}
-
-			// Get distinct organization_ids from this table
-			const orgResults = await sequelize.query(
-				`SELECT DISTINCT organization_id::text as org_id
-				 FROM ${tableName}
-				 WHERE organization_id IS NOT NULL`,
-				{ type: Sequelize.QueryTypes.SELECT }
-			)
-
-			// Check each organization_id against CSV data
-			for (const row of orgResults) {
-				const orgId = row.org_id
-				if (!csvOrgIds.has(orgId)) {
-					missingOrgs.add(orgId)
-				}
-			}
-
-			console.log(`✅ Checked ${tableName}: ${orgResults.length} distinct organization_ids`)
-		} catch (error) {
-			console.warn(`⚠️  Failed to check ${tableName}: ${error.message}`)
-		}
-	}
-
-	// Also check user_id based tables via user_extensions
-	const userIdTables = [
-		'connection_requests',
-		'connections',
-		'entities',
-		'feedbacks',
-		'issues',
-		'modules',
-		'post_session_details',
-		'question_sets',
-		'questions',
-		'session_attendees',
-		'session_request',
-	]
-
-	for (const tableName of userIdTables) {
-		try {
-			// Check if table exists first
-			const tableExists = await sequelize.query(
-				`SELECT EXISTS (SELECT FROM information_schema.tables WHERE table_name = '${tableName}')`,
-				{ type: Sequelize.QueryTypes.SELECT }
-			)
-
-			if (!tableExists[0].exists) {
-				console.log(`⚠️  Table ${tableName} does not exist, skipping`)
-				continue
-			}
-
-			// Get organization_ids via user_extensions
-			const orgResults = await sequelize.query(
-				`SELECT DISTINCT ue.organization_id::text as org_id
-				 FROM user_extensions ue
-				 INNER JOIN ${tableName} t ON t.user_id = ue.user_id
-				 WHERE ue.organization_id IS NOT NULL`,
-				{ type: Sequelize.QueryTypes.SELECT }
-			)
-
-			// Check each organization_id against CSV data
-			for (const row of orgResults) {
-				const orgId = row.org_id
-				if (!csvOrgIds.has(orgId)) {
-					missingOrgs.add(orgId)
-				}
-			}
-
-			console.log(`✅ Checked ${tableName} (via user_extensions): ${orgResults.length} distinct organization_ids`)
-		} catch (error) {
-			console.warn(`⚠️  Failed to check ${tableName}: ${error.message}`)
-		}
-	}
-
-	// Report results
-	if (missingOrgs.size > 0) {
-		const missingOrgsList = Array.from(missingOrgs).sort()
-		console.error('\n❌ VALIDATION FAILED: Missing organization_ids in CSV')
-		console.error('='.repeat(60))
-		console.error(`Found ${missingOrgs.size} organization_ids in database that are missing from CSV:`)
-		missingOrgsList.forEach((orgId) => {
-			console.error(`   - organization_id: ${orgId}`)
-		})
-		console.error('\n📝 Required action:')
-		console.error(
-			'   - Add missing organization_ids to data_codes.csv with proper tenant_code and organization_code'
-		)
-		console.error('   - Or verify if these organization_ids should be removed from database')
-
-		throw new Error(
-			`Migration cannot proceed: ${missingOrgs.size} organization_ids missing from CSV. See details above.`
-		)
-	}
-
-	console.log('✅ Validation passed: All database organization_ids are covered in CSV')
-}
 
 console.log('🎯 Production Mentoring Service Data Migration')
 console.log('==============================================')
@@ -229,10 +90,6 @@ rl.question('\n🤔 Proceed with production migration? (y/N): ', async (answer) 
 			console.log(`   Batch size: ${migrator.batchSize}`)
 			console.log(`   Max retries: ${migrator.maxRetries}`)
 			console.log(`   Progress updates: Every ${migrator.progressInterval} records`)
-
-			// Load CSV data and validate before migration
-			await migrator.loadLookupData()
-			await validateDatabaseOrgsCoveredByCSV(migrator.orgLookupCache, migrator.sequelize)
 
 			await migrator.execute()
